+++ index.html
@@ -5,11 +5,7 @@
   <meta name="keywords" content="poker">
   <meta name="description" content="This is a simplified poker">
   <title>poker</title>
-  <link rel="stylesheet" href="styles/index.css">
-  <link rel="stylesheet" href="styles/bootstrap.min.css">
-  <script type="text/javascript" src="scripts/av-min.js"></script>
-  <script type="text/javascript" src="scripts/jquery.min.js"></script>
-  <script type="text/javascript" src="scripts/require.js" data-main="scripts/index"></script>
+  <link rel="stylesheet" href="dist/bundle.css">
 </head>
 <body>
   <section class="welcome_page in_center">
@@ -20,4 +16,7 @@
     <br />
     <a href="session_new.html"> New Session </a>
   </section>
+  <script type="text/javascript" src="dist/vendor.js"></script>
+  <script type="text/javascript" src="dist/index.js"></script>
 </body>
+</html>
