+++ join_session.html
@@ -5,11 +5,7 @@
   <meta name="keywords" content="poker">
   <meta name="description" content="This is a simplified poker">
   <title>poker</title>
-  <link rel="stylesheet" href="styles/index.css">
-  <link rel="stylesheet" href="styles/bootstrap.min.css">
-  <script type="text/javascript" src="scripts/av-min.js"></script>
-  <script type="text/javascript" src="scripts/jquery.min.js"></script>
-  <script type="text/javascript" src="scripts/require.js" data-main="scripts/join_session"></script>
+  <link rel="stylesheet" href="dist/bundle.css">
 </head>
 <body>
   <section class="join_session_page in_center">
@@ -18,4 +14,7 @@
       <button class="btn btn-primary" id="join_with_name"> Join </button>
     </form>
   </section>
+  <script type="text/javascript" src="dist/vendor.js"></script>
+  <script type="text/javascript" src="dist/join_session.js"></script>
 </body>
+</html>
