+++ session.html
@@ -5,12 +5,7 @@
   <meta name="keywords" content="poker">
   <meta name="description" content="This is a simplified poker">
   <title>poker</title>
-  <link rel="stylesheet" href="styles/index.css">
-  <link rel="stylesheet" href="styles/bootstrap.min.css">
-  <script type="text/javascript" src="scripts/av-min.js"></script>
-  <script type="text/javascript" src="scripts/jquery.min.js"></script>
-  <script type="text/javascript" src="scripts/js.cookie.js"></script>
-  <script type="text/javascript" src="scripts/require.js" data-main="scripts/session"></script>
+  <link rel="stylesheet" href="dist/bundle.css">
 </head>
 <body>
   <section id="session_page">
@@ -28,5 +23,7 @@
       </tr>
     </table>
   </section>
+  <script type="text/javascript" src="dist/vendor.js"></script
+  <script type="text/javascript" src="dist/session.js"></script>
 </body>
-
+</html>
