+++ session_new.html
@@ -5,12 +5,7 @@
   <meta name="keywords" content="poker">
   <meta name="description" content="This is a simplified poker">
   <title>poker</title>
-  <link rel="stylesheet" href="styles/index.css">
-  <link rel="stylesheet" href="styles/bootstrap.min.css">
-  <script type="text/javascript" src="scripts/av-min.js"></script>
-  <script type="text/javascript" src="scripts/jquery.min.js"></script>
-  <script type="text/javascript" src="scripts/require.js" data-main="scripts/index"></script>
-  <script type="text/javascript" src="scripts/index.js"></script>
+  <link rel="stylesheet" href="dist/bundle.css">
 </head>
 <body>
   <section class="in_center">
@@ -27,4 +22,7 @@
     </table>
     <button class="btn btn-primary" id="create_session"> Create Session </button>
   </section>
+  <script type="text/javascript" src="dist/vendor.js"></script>
+  <script type="text/javascript" src="dist/session_new.js"></script>
 </body>
+</html>
